Include check for prefix
A tree that is no prefix cannot be extended and a prefix-tree has no
non-null probability. Therefore check if a tree has stopped being a
prefix when extending and only then calculate the probability.
This could be improved by doing the check right after an extension and
not only when the non-prefix-tree is tried to be extended the next time
(which is how it is done right now).